+++ app/src/main/java/com/example/depository_system/informs/RukuRecordItemInform.java
@@ -29,7 +29,7 @@ public class RukuRecordItemInform {
 
     public String inboundTime;
 
-    public Integer number;
+    public double number;
     public String depositoryId;
     public String inboundIdentifier;
     public String inboundItemId;
